Move common code to WallpaperBaseViewController
